requirements.txt
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
jieba==0.42.1
gensim==4.3.2
scikit-learn==1.4.2
matplotlib==3.8.4
wordcloud==1.9.3
Pillow==10.3.0
jupyterlab==4.1.8
ipykernel==6.29.4
pyecharts==2.0.9
tqdm==4.66.4
nltk==3.8.1
joblib==1.3.2
xgboost==3.1.2